Moved Server and Client threads and handlers to new class/file.
