+++ Source/Cypher/app/src/main/java/com/semaphore_soft/apps/cypher/WiFiDirectBroadcastReceiver.java
@@ -9,20 +9,11 @@ import android.net.NetworkInfo;
 import android.net.wifi.WifiManager;
 import android.net.wifi.p2p.WifiP2pInfo;
 import android.net.wifi.p2p.WifiP2pManager;
-import android.os.AsyncTask;
-import android.os.Bundle;
-import android.os.Handler;
-import android.os.Message;
 import android.util.Log;
 import android.view.View;
 import android.widget.Toast;
 
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
 import java.net.InetAddress;
-import java.net.ServerSocket;
-import java.net.Socket;
 import java.net.UnknownHostException;
 
 /**
@@ -37,8 +28,6 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver
     private final WifiP2pManager.Channel mChannel;
     private final MainActivity mActivity;
     private boolean connecting = false;
-
-    public static Socket mySocket = null;
 
     private final static String TAG = "WifiBR";
 
@@ -69,8 +58,10 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver
                             // ex. Create server thread and listen for incoming connections
                             Log.d(TAG, "Device is group owner");
                             Toast.makeText(mActivity, "You're the group owner!", Toast.LENGTH_SHORT).show();
-                            new Thread(new ServerThread()).start();
                             mActivity.stopServices();
+                            // ServerThread is not static so it requires an instance of the outer class
+                            new Thread(new DeviceThreads(mActivity).new ServerThread()).start();
+
                         }
                         else if (wifiP2pInfo.groupFormed)
                         {
@@ -78,10 +69,9 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver
                             // Create client thread that connects to group owner
                             Log.d(TAG, "Device is in group");
                             Toast.makeText(mActivity, "You're in a group!", Toast.LENGTH_SHORT).show();
-                            //new Thread(new ClientThread(groupOwnerAddress)).start();
                             mActivity.stopServices();
-                            new Client().execute(groupOwnerAddress);
-
+                            // ClientThread is not static so it requires an instance of the outer class
+                            new Thread(new DeviceThreads(mActivity).new ClientThread(groupOwnerAddress)).start();
                         }
                     }
                     catch (UnknownHostException e)
@@ -179,207 +169,7 @@ public class WiFiDirectBroadcastReceiver extends BroadcastReceiver
         }
     }
 
-    private Handler handler = new Handler(new Handler.Callback()
-    {
-        @Override
-        public boolean handleMessage(Message msg)
-        {
-            mActivity.setLabel(msg.getData().getString("msg"));
-            return true;
-        }
-    });
-    public void mkmsg(String str)
-    {
-        Message msg = new Message();
-        Bundle b = new Bundle();
-        b.putString("msg", str);
-        msg.setData(b);
-        handler.sendMessage(msg);
-    }
 
-    // Handler to get toasts for debugging
-    private Handler tHandler = new Handler(new Handler.Callback()
-    {
-        @Override
-        public boolean handleMessage(Message msg)
-        {
-            mActivity.toasts(msg.getData().getString("msg"));
-            return true;
-        }
-    });
-    public void makeToast(String str)
-    {
-        Message msg = new Message();
-        Bundle b = new Bundle();
-        b.putString("msg", str);
-        msg.setData(b);
-        tHandler.sendMessage(msg);
-    }
-
-    public static void write(String str)
-    {
-        if (mySocket != null)
-        {
-            Log.d("Write", "sending message");
-            try
-            {
-                DataOutputStream dos = new DataOutputStream(mySocket.getOutputStream());
-                dos.writeUTF(str);
-                dos.flush();
-            } catch (IOException e)
-            {
-                e.printStackTrace();
-            }
-        }
-        else
-        {
-            Log.d("Write", "Socket is null");
-        }
-    }
-
-    private class ServerThread extends Thread
-    {
-        // The local server socket
-        private ServerSocket serverSocket = null;
-        private Socket my_socket;
-
-        public ServerThread()
-        {
-            try
-            {
-                serverSocket = new ServerSocket(MainActivity.SERVER_PORT);
-
-            }
-            catch (IOException e)
-            {
-                e.printStackTrace();
-                Log.e("ServerThread", "Failed to start server");
-                Toast.makeText(mActivity, "Failed to start server", Toast.LENGTH_SHORT).show();
-            }
-        }
-
-        public void run()
-        {
-            Log.i("ServerThread", "Waiting on accept");
-            makeToast("Waiting on accept");
-            my_socket = null;
-            try
-            {
-                my_socket = serverSocket.accept();
-            }
-            catch (IOException e)
-            {
-                e.printStackTrace();
-                Log.e("ServerThread", "Failed to accept connection");
-                makeToast("Failed to accept connection");
-            }
-            if (my_socket != null)
-            {
-                Log.i("ServerConnect", "Connection made");
-                makeToast("Connection made");
-                try
-                {
-                    DataOutputStream out = new DataOutputStream(my_socket.getOutputStream());
-                    // This may or may not be needed
-                    out.flush();
-                    DataInputStream in = new DataInputStream((my_socket.getInputStream()));
-                    out.writeUTF("Hello, World!");
-                    // flush after write or inputStream will hang on read
-                    out.flush();
-                    Log.d("ServerThread", "sent message");
-                    mkmsg(in.readUTF());
-                    Log.d("ServerThread", "received message");
-                    out.close();
-                    in.close();
-                    my_socket.close();
-                }
-                catch (IOException e)
-                {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-//    private class ClientThread extends Thread
-//    {
-//        private Socket my_socket = null;
-//
-//        public ClientThread(InetAddress address)
-//        {
-//            try
-//            {
-//                // creates and connects to address at specified port
-//                my_socket = new Socket(address, MainActivity.SERVER_PORT);
-//            }
-//            catch (IOException e)
-//            {
-//                e.printStackTrace();
-//                Log.e("ClientThread", "Failed to start socket");
-//                makeToast("Failed to start socket");
-//            }
-//        }
-//
-//        public void run()
-//        {
-//            // Connection was accepted
-//            if (my_socket != null)
-//            {
-//                Log.i("ClientThread", "Connection made");
-//                try
-//                {
-//                    DataOutputStream out = new DataOutputStream(my_socket.getOutputStream());
-//                    out.flush();
-//                    DataInputStream in = new DataInputStream(my_socket.getInputStream());
-//                    mkmsg(in.readUTF());
-//                    out.writeUTF("Hello, World!");
-//                    // flush after write or inputStream will hang on read
-//                    out.flush();
-//                    Log.d("ClientThread", "sent message");
-//                    out.close();
-//                    in.close();
-//                    my_socket.close();
-//                }
-//                catch (IOException e)
-//                {
-//                    e.printStackTrace();
-//                }
-//            }
-//        }
-//
-//    }
-
-    private class Client extends AsyncTask<InetAddress, DataOutputStream, String>
-    {
-        @Override
-        protected String doInBackground(InetAddress... params)
-        {
-            try
-            {
-                mySocket = new Socket(params[0], MainActivity.SERVER_PORT);
-                Log.d("Client", "Connection made");
-                DataOutputStream out = new DataOutputStream(mySocket.getOutputStream());
-                out.flush();
-                DataInputStream in = new DataInputStream(mySocket.getInputStream());
-                mkmsg(in.readUTF());
-                //publishProgress(out);
-            }
-            catch (IOException e)
-            {
-                e.printStackTrace();
-                Log.e("Client", "Connection failed");
-            }
-
-            return "";
-        }
-
-        @Override
-        protected void onProgressUpdate(DataOutputStream... progress)
-        {
-            write("Hello, World!");
-            Log.d("Client", "Sending message");
-        }
-    }
 }
